Rename: myresult -> chartBarView로 변경
