+++ src/components/ChartBarView.vue
